airplay: Rename auth to auth_legacy
Relates to #1255
